Refactor: move getLocalIpAddress method to controller and add isServerReachable method for join handling
